Use newest version of getSASParams
This means that instead of passing in filterConfig, we convert it to
the filterKeys format and pass that in instead. The simple and
reusable filterConfig2filterKeys function does the work.

This is a bit around-the-houses, but it does have the benefit that we
no longer need to specify each filter value twice, once as its name
and once as its CQL value.

And of course it means that we can dump our own copy of getSASParams
completely once that function is exposed in a release of the stripes
platform.