Subtler configuration of react/jsx-one-expression-per-line
In fact, this seems to behave exactly the same as disabling the rule
altogether -- as does literally every other configuration of it that
I've tried, including `"react/jsx-one-expression-per-line": "on"`. But
at least now it's saying what I mean. If ESLint doesn't interpret it
right, that can be a problem for another day.