First functional version of status settings
Works for listing statuses. But create, edit and delete all fail for
reasons to do with the back-end. They may start working on their own
when the back-end is fixed, or it may turn out that I have
misunderstood the intention of the back-end and thereis more work to
do here.