Fetch 100 tags, not the default of 10
This is because the `directory/tags` endpoint returns many many
duplictes, so that by the time we de-dup the first ten records, we
find we've only seen one tag ("Branch") over and over -- rendering the
tags filter useless.

By fetching 100 we have a better chance of getting a representative
sample, and indeed we now see five tags (Branch, CONSORTIUM,
INSTITUTION, Main, and Storage).

It would better to fetch even more tags, but I can't get it to fetch
more than 100 at at a time, not even if I also set `perRequest` to
(say) 1000. I am guessing there is a hard limit somewhere, and if we
really want more than 100 tags we'll have to do paging. Let's not.

Of course, the real solution would be for the `directory/tags`
endpoint not to give us duplicates at all!