First stab at automatic update
It "works" in that it does refresh every five seconds, and preserves
the query state as it does so. But the refresh is big and ugly and
slow.

I think I can ameliorate that by not having it render the
"... Loading" caption. But that won’t fix the more fundamental problem
that it scrolls to the top every time.

My question: why doesn’t React notice that the new props are the same
as the old, and skip the re-render?

My tentative answer: because it passes through a state where the props
are different, because stripes-connect is passing the resource.dirents
prop with status=loading.

And I really can’t see a good way to avoid that.