Add "(DO NOT USE)" to the ends of permission names
_sigh_

Like an idiot, when I added UI-level directory permissions in commits
057700b0acc7090694f96f0111ffc87516f6b439
and
357537e84d2c7280efbd8da078d176611843ec70
I gave them names that begin with "directory" instead of "ui-directory".

I could just rename them, but then the permissions would would persist
under their old names in running systems as well as under the new
ones, and there is no way to remove a permission once it exists.

So stage 1 is that I have modified the human-readable names of the
permissions in question to include "(DO NOT USE)" and the end. Once
the system has been rebuilt, the permission names will appear in that
form.

Then I will execute stage 2, which is to remove those suffixes from
the human-readable permission names, and add the mising "ui-" prefixes
to the machine-readable permission names. Once the system has been
re-rebuilt, this will mean each permission appears twice in the list:
once with the old machine-readable name and with "(DO NOT USE)" on the
end of its human-readable name; and once with the new machine-readable
name and the correct human-readable name.

Stage three will be the next time the system is rebuilt from the
ground up: the old "(DO NOT USE)" permissions will then be omitted.