Edit form correctly handles "ui-directory.edit-local" permission
When the user has this permission but not `ui-directory.edit-all`, the
**Edit** button is enabled in the **Actions** when viewing a record,
but clicking through that button takes you to a version of the edit
form in which only the **Local information** tab is present. In
addition, it starts that tab with the **Local directory entry info**
accordion open, because why would you ever not want it to be?